requirements.txt
dash>=2.17
dash-cytoscape>=0.3.0
pandas>=2.2
numpy>=1.26
pydantic>=2.7
python-dateutil>=2.9
Flask-Caching>=2.1
gunicorn>=21.2
plotly>=5.22
pytest>=8.2
mypy>=1.10
ruff>=0.5
python-dotenv>=1.0
openpyxl>=3.1
reportlab>=3.6